+++ sct_sentinel1_reader/src/sct_sentinel1_reader/__init__.py
@@ -10,4 +10,4 @@ from sct_sentinel1_reader.interface import get_ale_corrector, get_detector, get_
 
 __all__ = ["get_manager", "get_detector", "get_ale_corrector"]
 
-__version__ = "1.1.2"
+__version__ = "1.1.3"